# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TalentMesh API",
    "version": "0.1.0",
    "description": "Generated from app/api/ by scripts/generate-openapi.ts — do not edit by hand, run `npm run docs:openapi`.\n\nAuth is a session cookie (`tm_access_token`), not a bearer header. Roles and permissions shown as `x-allowed-roles` / `x-required-permission` are read from the withApi() options that actually enforce them.\n\nOperations flagged `x-undocumented-body` do not use withApi(), so their request shape is not declared anywhere machine-readable and has been left blank rather than guessed."
  },
  "servers": [
    {
      "url": "https://talentmeshsolutions.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tm_access_token"
      }
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/auth/email/{slug}": {
      "post": {
        "operationId": "post_api_auth_email_slug",
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-undocumented-body": "Route does not use withApi(); request shape is not declared in a Zod schema and was not inferred.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Unauthorized — no valid session cookie"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "post_api_auth_logout",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-undocumented-body": "Route does not use withApi(); request shape is not declared in a Zod schema and was not inferred.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Unauthorized — no valid session cookie"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/auth/oauth/exchange": {
      "post": {
        "operationId": "post_api_auth_oauth_exchange",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-undocumented-body": "Route does not use withApi(); request shape is not declared in a Zod schema and was not inferred.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Unauthorized — no valid session cookie"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}